+++ packages/playground/src/data/actions/setSpeechSynthesisDeploymentId.js
@@ -1,0 +1,10 @@
+const SET_SPEECH_SYNTHESIS_DEPLOYMENT_ID = 'SET_SPEECH_SYNTHESIS_DEPLOYMENT_ID';
+
+export default function (deploymentId) {
+  return {
+    type: SET_SPEECH_SYNTHESIS_DEPLOYMENT_ID,
+    payload: { deploymentId }
+  };
+}
+
+export { SET_SPEECH_SYNTHESIS_DEPLOYMENT_ID }