+++ packages/playground/src/data/sagas/setPonyfill.js
@@ -10,10 +10,12 @@ import { SET_BING_SPEECH_SUBSCRIPTION_KEY } from '../actions/setBingSpeechSubscr
 import { SET_ON_DEMAND_AUTHORIZATION_TOKEN } from '../actions/setOnDemandAuthorizationToken';
 import { SET_PONYFILL_TYPE } from '../actions/setPonyfillType';
 import { SET_REGION } from '../actions/setRegion';
+import { SET_SPEECH_RECOGNITION_ENDPOINT_ID } from '../actions/setSpeechRecognitionEndpointId';
 import { SET_SPEECH_RECOGNITION_REFERENCE_GRAMMARS } from '../actions/setSpeechRecognitionReferenceGrammars';
 import { SET_SPEECH_RECOGNITION_TEXT_NORMALIZATION } from '../actions/setSpeechRecognitionTextNormalization';
 import { SET_SPEECH_SERVICES_AUTHORIZATION_TOKEN } from '../actions/setSpeechServicesAuthorizationToken';
 import { SET_SPEECH_SERVICES_SUBSCRIPTION_KEY } from '../actions/setSpeechServicesSubscriptionKey';
+import { SET_SPEECH_SYNTHESIS_DEPLOYMENT_ID } from '../actions/setSpeechSynthesisDeploymentId';
 import setPonyfill from '../actions/setPonyfill';
 
 import createBingSpeechPonyfill, { fetchAuthorizationToken as fetchBingSpeechAuthorizationToken } from 'web-speech-cognitive-services/lib/BingSpeech';
@@ -29,9 +31,11 @@ export default function* () {
       || type === SET_PONYFILL_TYPE
       || type === SET_REGION
       || type === SET_SPEECH_RECOGNITION_REFERENCE_GRAMMARS
+      || type === SET_SPEECH_RECOGNITION_ENDPOINT_ID
       || type === SET_SPEECH_RECOGNITION_TEXT_NORMALIZATION
       || type === SET_SPEECH_SERVICES_AUTHORIZATION_TOKEN
       || type === SET_SPEECH_SERVICES_SUBSCRIPTION_KEY
+      || type === SET_SPEECH_SYNTHESIS_DEPLOYMENT_ID
       || type === SET_ON_DEMAND_AUTHORIZATION_TOKEN,
     setPonyfillSaga
   );
@@ -44,10 +48,12 @@ function* setPonyfillSaga() {
     onDemandAuthorizationToken,
     ponyfillType,
     region,
+    speechRecognitionEndpointId,
     speechRecognitionReferenceGrammars: referenceGrammars,
     speechRecognitionTextNormalization: textNormalization,
     speechServicesAuthorizationToken,
-    speechServicesSubscriptionKey
+    speechServicesSubscriptionKey,
+    speechSynthesisDeploymentId
   } = yield select();
 
   if (ponyfillType === 'browser') {
@@ -69,7 +75,11 @@ function* setPonyfillSaga() {
           authorizationToken: onDemandAuthorizationToken ? () => {
             console.log('On-demand fetching Bing Speech authorization token');
 
-            return fetchBingSpeechAuthorizationToken(bingSpeechSubscriptionKey);
+            try {
+              return fetchBingSpeechAuthorizationToken(bingSpeechSubscriptionKey);
+            } catch (err) {
+              console.error('Failed to fetch Bing Speech authorization token', err);
+            }
           } : null,
           subscriptionKey: onDemandAuthorizationToken ? null : bingSpeechSubscriptionKey
         }
@@ -80,6 +90,8 @@ function* setPonyfillSaga() {
     const options = {
       referenceGrammars,
       region,
+      speechRecognitionEndpointId,
+      speechSynthesisDeploymentId,
       textNormalization
     };
 
@@ -93,10 +105,14 @@ function* setPonyfillSaga() {
           authorizationToken: onDemandAuthorizationToken ? () => {
             console.log('On-demand fetching Speech Services authorization token');
 
-            return fetchSpeechServicesAuthorizationToken({
-              region,
-              subscriptionKey: speechServicesSubscriptionKey
-            });
+            try {
+              return fetchSpeechServicesAuthorizationToken({
+                region,
+                subscriptionKey: speechServicesSubscriptionKey
+              });
+            } catch (err) {
+              console.error('Failed to fetch Speech Services authorization token', err);
+            }
           } : null,
           subscriptionKey: onDemandAuthorizationToken ? null : speechServicesSubscriptionKey
         }